integrate ProtonGE for steam (flatpak version)

+++ wannaplay.sh
@@ -304,18 +304,33 @@ fi
 
 #Install ProtonGE Custom Build
 
+
+
+
 instprotonGE() {
-    #FIXME: protonGE get not listet in steam [Game Breaker]
-    echo -e ${GREEN}"TASK: Installing Proton-GE Custom Build for native Steam"${NC}
-    jqcheck
-    rm /tmp/Proton*
-    protonGElink=$(wget -q -nv -O- https://api.github.com/repos/GloriousEggroll/proton-ge-custom/releases/latest 2>/dev/null |  jq -r '.assets[] | select(.browser_download_url | contains("Proton")) | .browser_download_url')
+
+steamapt=$(apt list steam --installed 2>/dev/null | grep -ow "steam")
+steamflat=$(flatpak list --app | grep -ow 'Steam' | tail -1)
+
+if [ -z "$steamapt" ]; then
+
     mkdir -p /home/$real_user/.steam/root/compatibilitytools.d
     wget $protonGElink -P /tmp/
     tar xf /tmp/Proton*.tar.gz -C /home/$real_user/.steam/root/compatibilitytools.d
     chown -R $real_user:$real_user /home/$real_user/.steam
     rm /tmp/Proton*
 
+
+	if [ -z "$steamflat" ]; then
+
+    mkdir /var/lib/flatpak/app/com.valvesoftware.Steam/x86_64/stable/17c6c458a53f9dd0f7661438be8bc2f5a7eaae73a42a49a4bd4778bd2f3e623b/files/share/steam/compatibilitytools.d
+    wget $protonGElink -P /tmp/
+    tar xf /tmp/Proton*.tar.gz -C /var/lib/flatpak/app/com.valvesoftware.Steam/x86_64/stable/17c6c458a53f9dd0f7661438be8bc2f5a7eaae73a42a49a4bd4778bd2f3e623b/files/share/steam/compatibilitytools.d
+    #chown -R $real_user:$real_user /var/lib/flatpak/app/com.valvesoftware.Steam/x86_64/stable/17c6c458a53f9dd0f7661438be8bc2f5a7eaae73a42a49a4bd4778bd2f3e623b/files/share/steam/compatibilitytools.d
+    rm /tmp/Proton*
+
+    fi
+ fi
 }
 
 #install Protontricks + GUI
@@ -566,6 +581,10 @@ if [ $instprotonGE == "true" ]; then
 
     echo -e ${GREEN}"TASK: <Installing latest ProtonGE Release"${NC}
     jqcheck
+    echo -e ${GREEN}"TASK: Installing Proton-GE Custom Build for native Steam"${NC}
+    jqcheck
+    rm /tmp/Proton*
+    protonGElink=$(wget -q -nv -O- https://api.github.com/repos/GloriousEggroll/proton-ge-custom/releases/latest 2>/dev/null |  jq -r '.assets[] | select(.browser_download_url | contains("Proton")) | .browser_download_url')
     instprotonGE
        
 fi
